COMP: Fix -Wsign-compare warnings in SurfaceFeaturesExtractor filter and CLI

+++ CLI/SurfaceFeaturesExtractor/surfacefeaturesextractor.hxx
@@ -98,7 +98,7 @@ void SurfaceFeaturesExtractor::compute_distances()
 
 	// Load each mean groupe shape & create labels
 	std::vector<std::string> meanDistLabels;
-	for (int k=0; k<this->meanShapesList.size(); k++) 
+	for (size_t k=0; k<this->meanShapesList.size(); k++)
 	{
 		std::ostringstream k_temp;
     	k_temp << k;
@@ -109,7 +109,7 @@ void SurfaceFeaturesExtractor::compute_distances()
 	//for (int k=0; k<this->meanShapesList.size(); k++) 
 	//	std::cout<<meanDistLabels[k]<<std::endl;
 
-	for(int k=0; k<meanShapesList.size(); k++)
+	for(size_t k=0; k<meanShapesList.size(); k++)
 	{
 		vtkSmartPointer<vtkFloatArray> meanDistance = vtkSmartPointer<vtkFloatArray>::New() ;
 		meanDistance->SetName(meanDistLabels[k].c_str());
